Redesign Phase 2: Part 2.
This is part 2 of the redesign, it rolls back to the old design on some pages while also changing the layout to something more similar to the previous one. This update also adds links to GNU on the login page and removes the TOS though the TOS page is still available.